feat: Add chemical filter tooltip display and rearrange base client
