+++ enderio-modded-conduits/src/main/java/com/enderio/modded_conduits/client/modules/mekanism/MekanismClientModule.java
@@ -1,13 +1,14 @@
 package com.enderio.modded_conduits.client.modules.mekanism;
 
-import com.enderio.enderio.api.conduits.model.RegisterConduitModelModifiersEvent;
 import com.enderio.enderio.api.conduits.screen.RegisterConduitScreenTypesEvent;
 import com.enderio.modded_conduits.client.ConduitClientModule;
 import com.enderio.modded_conduits.client.modules.mekanism.screens.ChemicalConduitScreenType;
 import com.enderio.modded_conduits.client.modules.mekanism.screens.EnderChemicalFilterScreen;
 import com.enderio.modded_conduits.client.modules.mekanism.screens.HeatConduitScreenType;
 import com.enderio.modded_conduits.common.modules.mekanism.MekanismModule;
+import com.enderio.modded_conduits.common.modules.mekanism.chemical_filter.EnderChemicalFilter;
 import net.neoforged.bus.api.IEventBus;
+import net.neoforged.neoforge.client.event.RegisterClientTooltipComponentFactoriesEvent;
 import net.neoforged.neoforge.client.event.RegisterMenuScreensEvent;
 
 public class MekanismClientModule implements ConduitClientModule {
@@ -21,6 +22,7 @@ public class MekanismClientModule implements ConduitClientModule {
     public void initialize(IEventBus modEventBus) {
         modEventBus.addListener(this::registerConduitScreenTypes);
         modEventBus.addListener(this::registerMenuScreens);
+        modEventBus.addListener(this::registerClientComponents);
     }
 
     private void registerConduitScreenTypes(RegisterConduitScreenTypesEvent event) {
@@ -30,5 +32,9 @@ public class MekanismClientModule implements ConduitClientModule {
 
     private void registerMenuScreens(RegisterMenuScreensEvent event) {
         event.register(MekanismModule.CHEMICAL_FILTER_MENU.get(), EnderChemicalFilterScreen::new);
+    }
+
+    private void registerClientComponents(RegisterClientTooltipComponentFactoriesEvent event) {
+        event.register(EnderChemicalFilter.class, ClientEnderChemicalFilterTooltip::new);
     }
 }
