+++ enderio-modded-conduits/src/main/java/com/enderio/modded_conduits/common/modules/mekanism/chemical_filter/EnderChemicalFilter.java
@@ -10,13 +10,14 @@ import net.minecraft.core.NonNullList;
 import net.minecraft.network.RegistryFriendlyByteBuf;
 import net.minecraft.network.codec.ByteBufCodecs;
 import net.minecraft.network.codec.StreamCodec;
+import net.minecraft.world.inventory.tooltip.TooltipComponent;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.List;
 import java.util.Objects;
 
 public record EnderChemicalFilter(NonNullList<ChemicalStack> matches, boolean isDenyList)
-    implements ChemicalFilter {
+    implements ChemicalFilter, TooltipComponent {
 
     public static final EnderChemicalFilter EMPTY = new EnderChemicalFilter(0);
