+++ enderio-modded-conduits/src/main/java/com/enderio/modded_conduits/common/modules/mekanism/chemical_filter/EnderChemicalFilterItem.java
@@ -3,17 +3,21 @@ package com.enderio.modded_conduits.common.modules.mekanism.chemical_filter;
 import com.enderio.core.annotations.UseOnly;
 import com.enderio.enderio.content.filters.AbstractFilterItem;
 import com.enderio.enderio.content.filters.AbstractFilterMenu;
+import com.enderio.enderio.content.filters.FiltersLang;
 import com.enderio.modded_conduits.common.modules.mekanism.MekanismModule;
 import net.minecraft.core.component.DataComponentType;
 import net.minecraft.network.RegistryFriendlyByteBuf;
+import net.minecraft.network.chat.Component;
 import net.minecraft.world.entity.player.Inventory;
 import net.minecraft.world.inventory.AbstractContainerMenu;
 import net.minecraft.world.inventory.MenuType;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.TooltipFlag;
 import net.neoforged.fml.LogicalSide;
 import net.neoforged.neoforge.capabilities.ICapabilityProvider;
 import net.neoforged.neoforge.registries.DeferredHolder;
 
+import java.util.List;
 import java.util.function.Supplier;
 
 public class EnderChemicalFilterItem extends AbstractFilterItem<EnderChemicalFilter> {
@@ -41,6 +45,18 @@ public class EnderChemicalFilterItem extends AbstractFilterItem<EnderChemicalFil
     @Override
     protected AbstractContainerMenu createMenu(int containerId, Inventory playerInventory, AbstractFilterMenu.FilterAccess filterAccess) {
         return type.openMenu(containerId, playerInventory, filterAccess);
+    }
+
+    @Override
+    public void appendHoverText(ItemStack stack, TooltipContext context, List<Component> tooltipComponents, TooltipFlag tooltipFlag) {
+        super.appendHoverText(stack, context, tooltipComponents, tooltipFlag);
+
+        var filter = getFilter(stack);
+        if (filter.isDenyList()) {
+            tooltipComponents.add(FiltersLang.FILTER_DENY_LIST);
+        } else {
+            tooltipComponents.add(FiltersLang.FILTER_ALLOW_LIST);
+        }
     }
 
     public enum Type {
