+++ enderio/src/main/java/com/enderio/enderio/client/EnderIOClient.java
@@ -20,18 +20,18 @@ import com.enderio.enderio.client.content.conduits.model.bundle.ConduitBundleGeo
 import com.enderio.enderio.client.content.conduits.model.facades.FacadeItemGeometry;
 import com.enderio.enderio.client.content.conduits.model.modifier.RedstoneConduitModelModifier;
 import com.enderio.enderio.client.content.enderface.EnderfaceRenderer;
-import com.enderio.enderio.client.content.filters.ClientEnderFluidFilterTooltip;
-import com.enderio.enderio.client.content.filters.EnderFluidFilterScreen;
-import com.enderio.enderio.client.content.filters.EnderItemFilterScreen;
-import com.enderio.enderio.client.content.filters.LimitedItemFilterScreen;
-import com.enderio.enderio.client.content.filters.EnderSoulFilterScreen;
+import com.enderio.enderio.client.content.filters.fluid.ClientEnderFluidFilterTooltip;
+import com.enderio.enderio.client.content.filters.fluid.EnderFluidFilterScreen;
+import com.enderio.enderio.client.content.filters.item.EnderItemFilterScreen;
+import com.enderio.enderio.client.content.filters.item.LimitedItemFilterScreen;
+import com.enderio.enderio.client.content.filters.soul.EnderSoulFilterScreen;
 import com.enderio.enderio.client.content.filters.item.ClientEnderItemFilterTooltip;
 import com.enderio.enderio.client.content.filters.item.ClientExistingItemFilterTooltip;
 import com.enderio.enderio.client.content.filters.item.ClientLimitedItemFilterTooltip;
 import com.enderio.enderio.client.content.filters.redstone.RedstoneCountFilterScreen;
 import com.enderio.enderio.client.content.filters.redstone.RedstoneDoubleChannelFilterScreen;
 import com.enderio.enderio.client.content.filters.redstone.RedstoneTimerFilterScreen;
-import com.enderio.enderio.client.content.filters.ClientEnderSoulFilterTooltip;
+import com.enderio.enderio.client.content.filters.soul.ClientEnderSoulFilterTooltip;
 import com.enderio.enderio.client.content.fluid_tank.FluidTankBEWLR;
 import com.enderio.enderio.client.content.glass.GlassIconDecorator;
 import com.enderio.enderio.client.content.machines.IOOverlayBakedModel;
