+++ enderio/src/main/java/com/enderio/enderio/client/content/filters/item/DamageFilterModePickerWidget.java
@@ -1,4 +1,4 @@
-package com.enderio.enderio.client.content.filters;
+package com.enderio.enderio.client.content.filters.item;
 
 import com.enderio.core.client.gui.widgets.BaseEnumPickerWidget;
 import com.enderio.enderio.client.foundation.icon.EIOEnumIcons;
