+++ enderio/src/main/java/com/enderio/enderio/client/content/filters/item/package-info.java
@@ -1,0 +1,4 @@
+@javax.annotation.ParametersAreNonnullByDefault
+@net.minecraft.MethodsReturnNonnullByDefault
+
+package com.enderio.enderio.client.content.filters.item;
